feat: ScrapViewModel
